remove content that is not user related
We decided that the CodeCracker website will be focused
mostly on the users of CC, not on the programers.
We are moving all info related to building it to
the project itself. At this point it will mostly
be on the readme.md file. This might change in
the future.